misc/HtmlTestSuite/results/test-0164.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mailto:mstensho@chromium.org" rel="author" title="Morten Stenshorne"/>
<link href="https://www.w3.org/TR/css-page-3/#using-named-pages" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
<script>
  test_invalid_value("page", "not valid");
  test_invalid_value("page", "not,valid");
  test_invalid_value("page", "123px");
  test_invalid_value("page", "calc(10%+1px)");
  test_invalid_value("page", "default");
</script>
</head><body/></html>
